Refactor database access setup UI and update root route
Improves the database access setup flow with Turbo Frames and dynamic status handling in the UI, including auto-refresh for pending states and clearer error/retry messaging. Updates the root route to point to database access settings, cleans up JavaScript controller imports, and removes unused vendor stimulus-loading.js.
