+++ app/javascript/controllers/application.js
