+++ app/javascript/controllers/index.js
@@ -1,7 +1,4 @@
-// Import and register all your controllers from the importmaps under controllers/*
-
-import { application } from "./application"
-
-// Eager load all controllers defined in the import map under controllers/**/*_controller
+// Import and register all your controllers from the importmap via controllers/**/*_controller
+import { application } from "controllers/application"
 import { eagerLoadControllersFrom } from "@hotwired/stimulus-loading"
 eagerLoadControllersFrom("controllers", application)
