+++ config/importmap.rb
@@ -1,6 +1,7 @@
 # Pin npm packages by running ./bin/importmap
-pin "application"
-pin "@hotwired/turbo-rails", to: "https://cdn.jsdelivr.net/npm/@hotwired/turbo-rails@8.0.18/app/javascript/turbo/index.js"
-pin "@hotwired/stimulus", to: "https://cdn.jsdelivr.net/npm/@hotwired/stimulus@3.2.2/dist/stimulus.min.js"
-pin "@hotwired/stimulus-loading", to: "stimulus-loading.js"
+pin "application" # look in app/javascript/application.js or vendor/javascript/application.js
+pin "@hotwired/turbo-rails", to: "turbo.min.js" # imported in main application file: app/javascript/application.js
+pin "@hotwired/stimulus", to: "stimulus.min.js" # imported in app/javascript/application.js
+pin "@hotwired/stimulus-loading", to: "stimulus-loading.js" # imported in app/javascript/controllers/index.js
+
 pin_all_from "app/javascript/controllers", under: "controllers"
